Fix: update team api exception

+++ src/main/java/com/weave/weaveserver/service/TeamService.java
@@ -1,7 +1,5 @@
-
 package com.weave.weaveserver.service;
 
-import com.weave.weaveserver.config.exception.BadRequestException;
 import com.weave.weaveserver.config.exception.NotFoundException;
 import com.weave.weaveserver.config.jwt.TokenService;
 import com.weave.weaveserver.domain.Belong;
@@ -34,8 +32,10 @@ public class TeamService {
     private final UserRepository userRepository;
     private final BelongRepository belongRepository;
 
+
     @Transactional
-    public void createTeam(TeamRequest.createReq req, HttpServletRequest httpServletRequest) {
+    public ResponseEntity<JsonResponse> createTeam(TeamRequest.createReq req, HttpServletRequest httpServletRequest) {
+
         String userEmail = tokenService.getUserEmail(httpServletRequest);
         User leader = userRepository.findUserByEmail(userEmail);
 
@@ -49,17 +49,20 @@ public class TeamService {
                 .imgUrl(req.getImgUrl())
                 .isEmpty(true)
                 .build();
-        teamRepository.save(team);
+        Long teamIdx = teamRepository.save(team).getTeamIdx();
 
         Belong belong = Belong.builder()
                 .user(leader)
                 .team(team)
                 .build();
         belongRepository.save(belong);
+
+        return ResponseEntity.ok(new JsonResponse(200, "Success, createTeam",teamIdx));
+
     }
 
     @Transactional
-    public int addMember(Long teamIdx, TeamRequest.addMemberReq req, HttpServletRequest httpServletRequest) {
+    public ResponseEntity<JsonResponse> addMember(Long teamIdx, TeamRequest.addMemberReq req, HttpServletRequest httpServletRequest) {
 
         // 해당 팀 IDX 가 존재하지 않는 경우에 대한 예외 처리
         Team team = teamRepository.findById(teamIdx)
@@ -74,40 +77,44 @@ public class TeamService {
             //생성하려는 사용자와 팀짱이 동일
             System.out.println("same");
 
+            // find user (존재하는 사용자인지 확인)
+            User invitedUser = userRepository.findUserByEmail(req.getEmail());
+            if(invitedUser != null){
+                System.out.println("초대 팀원 존재 : "+ invitedUser.getEmail());
 
+                // 이미 팀에 존재하는 유저인지 확인
+                Belong already = belongRepository.findUserByIndex(teamIdx, invitedUser.getUserIdx());
+
+                if(already == null){
+                    System.out.println("팀짱이고 사용자도 존재하고, 팀에도 없어요~!!");
+
+                    Long userIdx = invitedUser.getUserIdx();
+                    User user = userRepository.getReferenceById(userIdx);
+
+                    Belong belong = Belong.builder()
+                            .user(user)
+                            .team(team)
+                            .build();
+
+                    belongRepository.save(belong);
+
+                    //team에 유저가 1명 이상 -> isEmpty = false;
+                    team.updateEmpty();
+
+                    return ResponseEntity.ok(new JsonResponse(200, "Success", userIdx));
+
+                } else {
+                    System.out.println("이미 초대된 팀원");
+                    return ResponseEntity.ok(new JsonResponse(2002, "이미 속한 팀원입니다.", null));
+                }
+            } else {
+                System.out.println("초대 팀원이 존재하지 않음");
+                return ResponseEntity.ok(new JsonResponse(2001, "해당 사용자가 존재하지 않습니다.", null));
+            }
         } else {
             //생성하려는 사용자와 팀짱이 동일하지 않음
             System.out.println("not same");
-            return 3000;
-        }
-
-
-        // find user (존재하는 사용자인지 확인)
-        User invitedUser = userRepository.findUserByEmail(req.getEmail());
-        if(invitedUser != null){
-            // 이미 팀에 존재하는 유저인지 확인
-            //System.out.println("이미 초대된 팀원입니다");
-
-            System.out.println("초대 팀원 존재 : "+ invitedUser.getEmail());
-
-            Long userIdx = invitedUser.getUserIdx();
-            User user = userRepository.getReferenceById(userIdx);
-
-            Belong belong = Belong.builder()
-                    .user(user)
-                    .team(team)
-                    .build();
-
-            belongRepository.save(belong);
-
-            //team에 유저가 1명 이상 -> isEmpty = false;
-            team.updateEmpty();
-
-            return 1;
-
-        } else {
-            System.out.println("초대 팀원이 존재하지 않음");
-            return 0;
+            return ResponseEntity.ok(new JsonResponse(2000, "권한이 없습니다. 팀장만 가능합니다.", null));
         }
     }
 
@@ -124,19 +131,18 @@ public class TeamService {
         List<TeamResponse.getMemberList> memberList = userList.stream().map(user -> new TeamResponse.getMemberList(
                 user.getUserIdx(),
                 user.getName(),
-                "user Image url"
+                user.getImage()
         )).collect(Collectors.toList());
 
         return memberList;
     }
 
     @Transactional
-    public List<TeamResponse.teamResponse> getMyTeams(HttpServletRequest httpServletRequest){
+    public ResponseEntity<JsonResponse> getMyTeams(HttpServletRequest httpServletRequest){
 
         String userEmail = tokenService.getUserEmail(httpServletRequest);
         User user = userRepository.findUserByEmail(userEmail);
-        //System.out.println(userEmail);
-        //System.out.println(user.getUserIdx());
+
         List<Team> teams = belongRepository.findAllByUserIdx(user.getUserIdx());
 
         List<TeamResponse.teamResponse> teamList = teams.stream().map(team -> new TeamResponse.teamResponse(
@@ -144,20 +150,26 @@ public class TeamService {
                 team.getTitle(),
                 team.getStartDate(),
                 team.getEndDate(),
-                "user Image url"
+                team.getImgUrl()
         )).collect(Collectors.toList());
 
-        return teamList;
+        if(teamList.isEmpty()){
+            return ResponseEntity.ok(new JsonResponse(2003, "속한 팀이 없습니다.", teamList));
+
+        } else {
+            return ResponseEntity.ok(new JsonResponse(200, "Success", teamList));
+        }
+
     }
 
     @Transactional
-    public void deleteTeam(Long teamIdx, HttpServletRequest httpServletRequest){
-        //팀짱인지 확인 과정 필요
+    public ResponseEntity<JsonResponse> deleteTeam(Long teamIdx, HttpServletRequest httpServletRequest){
 
         // 해당 팀 IDX 가 존재하지 않는 경우에 대한 예외 처리
         Team team = teamRepository.findById(teamIdx)
                 .orElseThrow(() -> new NotFoundException("해당 팀이 존재하지 않습니다."));
 
+        //팀짱인지 확인 과정 필요
         User leader = team.getLeader();
 
         String userEmail = tokenService.getUserEmail(httpServletRequest);
@@ -167,16 +179,17 @@ public class TeamService {
             //요청 사용자와 팀짱이 동일
             System.out.println("same");
             teamRepository.deleteByTeamIdx(teamIdx);
+            return ResponseEntity.ok(new JsonResponse(200, "Success", teamIdx));
 
         } else {
             //요청 사용자와 팀짱이 동일하지 않음
             System.out.println("not same");
+            return ResponseEntity.ok(new JsonResponse(2000, "권한이 없습니다. 팀장만 가능합니다.", null));
         }
-
     }
 
     @Transactional
-    public void deleteMember(Long teamIdx, TeamRequest.deleteMemberReq req, HttpServletRequest httpServletRequest){
+    public ResponseEntity<JsonResponse> deleteMember(Long teamIdx, TeamRequest.deleteMemberReq req, HttpServletRequest httpServletRequest){
 
         // 해당 팀 IDX 가 존재하지 않는 경우에 대한 예외 처리
         Team team = teamRepository.findById(teamIdx)
@@ -191,32 +204,44 @@ public class TeamService {
             //요청 사용자와 팀짱이 동일
             System.out.println("same");
 
+            // 삭제하려는 팀원이 존재하는 팀원이지 확인
+            User user = userRepository.findUserByEmail(req.getEmail());
 
+            if(user != null){
+                // 삭제하려는 팀원이 팀 소속인지 확인
+                Belong isBelong = belongRepository.findUserByIndex(teamIdx, user.getUserIdx());
 
-            // 삭제하려는 팀원이 팀 소속인지 확인
+                if(isBelong != null){
+                    // 소속이라면
+                    Belong deleteUser = belongRepository.findUserByIndex(teamIdx, user.getUserIdx());
+                    belongRepository.deleteById(deleteUser.getBelongIdx());
+                    return ResponseEntity.ok(new JsonResponse(200, "Success", user.getUserIdx()));
 
-
-
-            // 소속이라면
-            Belong deleteUser = belongRepository.findUserByIndex(teamIdx, req.getUserIdx());
-            belongRepository.deleteById(deleteUser.getBelongIdx());
+                } else {
+                    //소속이 아니라면
+                    System.out.println("팀 소속 유저가 아님");
+                    return ResponseEntity.ok(new JsonResponse(2001, "해당 사용자가 존재하지 않습니다.", null));
+                }
+            } else {
+                System.out.println("존재하지 않는 사용자");
+                return ResponseEntity.ok(new JsonResponse(2001, "해당 사용자가 존재하지 않습니다.", null));
+            }
 
         } else {
             //요청 사용자와 팀짱이 동일하지 않음
             System.out.println("not same");
+            return ResponseEntity.ok(new JsonResponse(2000, "권한이 없습니다. 팀장만 가능합니다.", null));
         }
-
-
     }
 
     @Transactional
-    public void updateTeam(Long teamIdx, TeamRequest.updateTeamReq req, HttpServletRequest httpServletRequest){
+    public ResponseEntity<JsonResponse> updateTeam(Long teamIdx, TeamRequest.updateTeamReq req, HttpServletRequest httpServletRequest){
+
+        // 해당 팀 IDX 가 존재하지 않는 경우에 대한 예외 처리
+        Team team = teamRepository.findById(teamIdx)
+                .orElseThrow(() -> new NotFoundException("해당 팀이 존재하지 않습니다."));
+
         //팀짱인지 확인 필요
-
-        // 해당 팀 IDX 가 존재하지 않는 경우에 대한 예외 처리
-        Team team = teamRepository.findById(teamIdx)
-                .orElseThrow(() -> new NotFoundException("해당 팀이 존재하지 않습니다."));
-
         User leader = team.getLeader();
 
         String userEmail = tokenService.getUserEmail(httpServletRequest);
@@ -225,13 +250,14 @@ public class TeamService {
         if(leader.equals(requester)){
             //요청 사용자와 팀짱이 동일
             System.out.println("same");
-
             //업데이트
             team.updateTeam(req.getTitle(), req.getStartDate(), req.getEndDate(), req.getImgUrl());
+            return ResponseEntity.ok(new JsonResponse(200, "Success", teamIdx));
 
         } else {
             //요청 사용자와 팀짱이 동일하지 않음
             System.out.println("not same");
+            return ResponseEntity.ok(new JsonResponse(2000, "권한이 없습니다. 팀장만 가능합니다.", null));
         }
     }
 }
